requirements.txt
tensorflow-gpu==1.10
scipy
scikit-learn
opencv-python
h5py
matplotlib
Pillow
requests
psutil
